refactor(unlock-app): Remove old checkout
