--- unlock-app/src/components/interface/checkout/Locks.tsx
@@ -1,90 +1,0 @@
-import React, { useState, useEffect } from 'react'
-import { useLock } from '../../../hooks/useLock'
-import { Lock } from './Lock'
-import { LoadingLock } from './LockVariations'
-
-interface LoadLockProps {
-  address: string
-  network: number
-  setHasKey: (key: any) => void
-  name: string
-  onSelected: (lock: any) => void
-}
-
-const LoadLock = ({
-  address,
-  network,
-  setHasKey,
-  name,
-  onSelected,
-}: LoadLockProps) => {
-  const [loading, setLoading] = useState(true)
-  const { lock, getLock } = useLock({ address }, network)
-
-  useEffect(() => {
-    const loadLock = async () => {
-      await getLock({ pricing: true })
-      setLoading(false)
-    }
-    loadLock()
-  }, [address])
-
-  if (loading) {
-    return <LoadingLock address={address} network={network} />
-  }
-  return (
-    <Lock
-      hasOptimisticKey={false}
-      purchasePending={false}
-      network={network}
-      lock={lock}
-      name={name}
-      setHasKey={setHasKey}
-      onSelected={onSelected}
-    />
-  )
-}
-interface LocksProps {
-  locks: any
-  network?: number
-  setHasKey: (key: any) => void
-  onSelected: (address: string) => void
-}
-
-interface LockProps {
-  network: number
-  name?: string
-}
-
-export const Locks = ({
-  network,
-  locks,
-  setHasKey,
-  onSelected,
-}: LocksProps) => {
-  return (
-    <div className="flex flex-col w-full gap-6">
-      {Object.entries(locks).map(
-        // @ts-expect-error
-        ([address, lockProps]: [string, LockProps]) => {
-          return (
-            <LoadLock
-              setHasKey={setHasKey}
-              // @ts-expect-error one of the two will be defined!
-
-              network={lockProps?.network || network}
-              key={address}
-              address={address}
-              name={lockProps?.name || ''}
-              onSelected={onSelected}
-            />
-          )
-        }
-      )}
-    </div>
-  )
-}
-
-Locks.defaultProps = {
-  network: 1,
-}
