+++ unlock-app/src/components/interface/checkout/index.tsx
@@ -5,7 +5,7 @@ import { useCheckoutCommunication } from '~/hooks/useCheckoutCommunication'
 import { getPaywallConfigFromQuery } from '~/utils/paywallConfig'
 import getOauthConfigFromQuery from '~/utils/oauth'
 import { useConfig } from '~/utils/withConfig'
-import { Checkout } from './Checkout'
+import { Checkout } from './main'
 import { Connect } from './Connect'
 import { Container } from './Container'
 import { CloseButton } from './Shell'
